renamed module and moved into init to save typing one level of hierarchy in the import process.
